+++ Client/src/pages/RecipesPage.jsx
@@ -1,0 +1,9 @@
+export default function Recipes() {
+    return (
+        <div>
+            <h1>Recipes</h1>
+            <h2>Recipes</h2>
+            <p>Recipes</p>
+        </div>
+    );
+};
