--- pages/RecipesPage.jsx
@@ -1,9 +1,0 @@
-export default function Recipes() {
-    return (
-        <div>
-            <h1>Recipes</h1>
-            <h2>Recipes</h2>
-            <p>Recipes</p>
-        </div>
-    );
-};
